support for define variable. support for special variables (e.g., tx-sender), working typecheck of names+tokens contract
